Fix deletion of plugin for unsupported PHP versions

+++ uninstall.php
@@ -12,14 +12,16 @@ defined( 'WP_UNINSTALL_PLUGIN' ) || die;
 
 require_once __DIR__ . '/site-reviews.php';
 
-$majorVersion = explode( '.', glsr_app()->version )[0];
+if( !glsr_version_check() )return;
+
+$majorVersion = explode( '.', glsr_app()->version );
 
 // remove plugin and widget options
-$options = [
-	sprintf( '%s-v%d', glsr_app()->prefix, $majorVersion ),
+$options = array(
+	sprintf( '%s-v%d', glsr_app()->prefix, array_shift( $majorVersion )),
 	sprintf( 'widget_%s_site-reviews', glsr_app()->id ),
 	sprintf( 'widget_%s_site-reviews-form', glsr_app()->id ),
-];
+);
 
 foreach( $options as $option ) {
 	delete_option( $option );
